fix: resolve CI failures — format, clippy, and pre-existing issues
Fix all CI check failures for the PR:

Format & Parse Gate:
- Run cargo fmt --all to fix formatting across entire workspace
- Fix duplicate import in rpc.rs doctest

Metrics Exporter CI:
- Fix cardinality_limit.rs clippy: add #[cfg(test)] to unused methods
- Fix unused variable `label_key` in cardinality_limit.rs
- Fix sse.rs unused field `contract_id` with #[allow(dead_code)]
- Fix pre-existing dead code warnings in rpc.rs with #[allow(dead_code)]

Pre-existing upstream fixes (needed for cargo check --workspace):
- Add missing deps to api-server/Cargo.toml (futures-util, utoipa, etc.)
- Fix tokio-tungstenite version (0.23.2 → 0.24)
- Fix prometheus proto API (get_name/get_value for 0.13)
- Add clippy allow attributes for contract crates
- Fix needless borrow in router-core
- Fix unnecessary cast in router-execution

Closes #198

+++ contracts/router-execution/src/lib.rs
@@ -557,11 +557,7 @@ impl RouterExecution {
             .get(&DataKey::ExecHistory)
             .unwrap_or(Vec::new(&env));
         let len = history.len();
-        let take = if limit as u32 > len {
-            len
-        } else {
-            limit as u32
-        };
+        let take = if limit > len { len } else { limit };
         let mut result = Vec::new(&env);
         // Return newest-first: iterate from the end
         let mut i = len;
@@ -910,7 +906,10 @@ mod tests {
         client.transfer_admin(&admin, &new_admin);
         let event = env.events().all().last().unwrap().clone();
         let topic: Symbol = event.1.get(0).unwrap().into_val(&env);
-        assert_eq!(topic, Symbol::new(&env, router_common::EVENT_ADMIN_TRANSFERRED));
+        assert_eq!(
+            topic,
+            Symbol::new(&env, router_common::EVENT_ADMIN_TRANSFERRED)
+        );
     }
 
     #[test]
